test/test262/tools/lint/requirements.txt
PyYAML==3.11
